# services/backend/src/main/java/com/fitnesscoaching/platform/modules/auth/adapter/in/web/dto/UserSettingsDto.java
package com.fitnesscoaching.platform.modules.auth.adapter.in.web.dto;

import java.util.Map;

public record UserSettingsDto(
        int weekStartsOn,
        String measurementSystem,
        Map<String, Object> accessibilityPreferences,
        Map<String, Object> privacyPreferences
) {
    public static UserSettingsDto defaults() {
        return new UserSettingsDto(1, "METRIC", Map.of(), Map.of());
    }
}